Fix build & test rule.
@andorp: you mentioned that it makes a difference whether the target
package is named here or not, but did not give any specifics.  one
specific difference is that if we name the package, no other packages
will be tested.  since in buildAll we now want to do just that, and
since this works fine for me, i am changing this back.