Remove dead code.
The "we're on the document page, but we don't have document" branch
can not be reached, and that is intentional.  The confusing thing
is perhaps that there is no click path back to the landing page,
but instead you have to reload the app.  But that's also intentional
becasue the landing page shouldn't really exist.  It's just a cheap
substitute for something else to come.