simplify VDoc OT types
This commit reduces the number of data types used in (OT.Edit RawContent).

Pros:
- less data type definitions is needed for Edit
- faster than the previous version

The tuning of the merge algorithm for edits was possible via adding more consturctors
to the data types which was deleted in this commit.
It is still possible to tune the merge algorithm by using special general data types
with custion Editable instances.