Fix: exception handling in runApp.
before this commit, exceptions were translated into an 'Either'
value by 'tryApp', and 'dbCommit' was run in a 'finally' clause
at the end of each 'AppM' transaction.

either of these two circumstances had the effect that 'dbCommit'
was called even after 'DBException's, and found an inconsistent
database state.  this caused the extremely unhelpful error message
"SqliteException (SQLite3 returned ErrorError while attempting
to perform step.)", and swallowed the actual error message.

in this commit, the error is twisted back into the ExceptT and
dbCommit is called *iff* ExceptT does not fire.